update to m95

+++ experimental/skrive/src/Artboard.cpp
@@ -43,7 +43,7 @@ std::tuple<sk_sp<Component>, size_t> parse_component(StreamReader* sr) {
     default: break;
   }
 
-  SkDebugf("!! unsupported node type: %d\n", block.type());
+  SkDebugf("!! unsupported node type: %d\n", (int)block.type());
   return {nullptr, 0};
 }
 
@@ -100,7 +100,7 @@ sk_sp<Artboard> parse_artboard(StreamReader* sr) {
 
     switch (block.type()) {
       case StreamReader::BlockType::kComponents: ab->setRoot(parse_components(sr)); break;
-      default: SkDebugf("!! Unsupported block type: %d\n", block.type()); break;
+      default: SkDebugf("!! Unsupported block type: %d\n", (int)block.type()); break;
     }
   }
