+++ src/gpu/vk/GrVkSampler.h
@@ -12,7 +12,9 @@
 #include "src/core/SkOpts.h"
 #include "src/gpu/vk/GrVkManagedResource.h"
 #include "src/gpu/vk/GrVkSamplerYcbcrConversion.h"
+
 #include <atomic>
+#include <cinttypes>
 
 class GrSamplerState;
 class GrVkGpu;
@@ -52,7 +54,7 @@ class GrVkSampler : public GrVkManagedResource {
 
 #ifdef SK_TRACE_MANAGED_RESOURCES
   void dumpInfo() const override {
-    SkDebugf("GrVkSampler: %d (%d refs)\n", fSampler, this->getRefCnt());
+    SkDebugf("GrVkSampler: %" PRIdPTR " (%d refs)\n", (intptr_t)fSampler, this->getRefCnt());
   }
 #endif
