+++ tools/trace/SkDebugfTracer.cpp
@@ -56,8 +56,8 @@ SkEventTracer::Handle SkDebugfTracer::addTraceEvent(
   if (open) {
     const char* category = this->getCategoryGroupName(categoryEnabledFlag);
     SkDebugf(
-        "[% 2d]%s <%s> %s%s #%d {\n", fIndent.size(), fIndent.c_str(), category, name, args.c_str(),
-        fCnt);
+        "[% 2d]%s <%s> %s%s #%d {\n", (int)fIndent.size(), fIndent.c_str(), category, name,
+        args.c_str(), fCnt);
     fIndent.append(" ");
   } else {
     SkDebugf("%s%s #%d\n", name, args.c_str(), fCnt);
@@ -69,5 +69,5 @@ SkEventTracer::Handle SkDebugfTracer::addTraceEvent(
 void SkDebugfTracer::updateTraceEventDuration(
     const uint8_t* categoryEnabledFlag, const char* name, SkEventTracer::Handle handle) {
   fIndent.resize(fIndent.size() - 1);
-  SkDebugf("[% 2d]%s } %s\n", fIndent.size(), fIndent.c_str(), name);
+  SkDebugf("[% 2d]%s } %s\n", (int)fIndent.size(), fIndent.c_str(), name);
 }
